code/GenAIEmailProcessingApp/requirements.txt
Flask==3.1.0
requests==2.32.3
google-genai==0.1.0
email==6.0.0a1
bootstrap-flask==2.4.2
llama-index-readers-file==0.4.6
unstructured==0.17.2
libmagic==1.0
apiflask==2.3.2
unstructured[all-docs]==0.17.2
libmagic==1.0
docx2txt==0.8
SQLAlchemy==2.0.39
flask_sqlalchemy==3.1.1
psycopg==3.2.6
psycopg[binary,pool]==3.2.6